pyproject.toml
[tool.poetry]
name = "pan-os-upgrade-scaffold"
version = "0.1.0"
description = "Scaffold for PAN-OS upgrade automation inspired by cdot65/pan-os-upgrade"
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "pan_os_upgrade" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
colorama = "^0.4.6"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"

[tool.poetry.scripts]
pan-os-upgrade = "pan_os_upgrade.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
